requirements.txt
cloudpickle==2.2.1
cookiecutter==2.1.1
decorator==5.1.1
docstring-to-markdown==0.12
faiss-cpu==1.7.4
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
gTTS==2.3.2
imagesize==1.4.1
langchain==0.0.222
langchainplus-sdk==0.0.20
matplotlib==3.7.1
matplotlib-inline==0.1.6
numexpr==2.8.4
numpy==1.24.4
numpydoc==1.5.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
pandas==2.0.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
regex==2023.6.3
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sklearn==0.0.post5
streamlit==1.24.0
streamlit-chat==0.1.1
tenacity==8.2.2
tiktoken==0.4.0
